requirements.txt
Flask
pandas
numpy
openpyxl
psycopg2-binary
sqlalchemy
python-dotenv
matplotlib
seaborn
plotly
google-generativeai
gunicorn
celery
xlsxwriter
dateparser
azure-storage-blob
reportlab
schedule
pytz
apscheduler
psycopg2-binary